pyproject.toml
[project]
name = "resumescreener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nltk>=3.9.1",
    "openai>=1.106.1",
    "pandas>=2.3.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "scikit-learn>=1.7.1",
    "tqdm>=4.67.1",
]
